+++ Rasbet/DTO/BetDTO/StatisticsDTO.cs
@@ -3,9 +3,9 @@
 public class StatisticsDTO
 {
     public int BetCount { get; set; }
-    public IDictionary<int, int> Statistics { get; set; }
+    public IDictionary<int, double> Statistics { get; set; }
 
-    public StatisticsDTO(int betCount, IDictionary<int, int> statistics)
+    public StatisticsDTO(int betCount, IDictionary<int, double> statistics)
     {
         BetCount = betCount;
         Statistics = statistics;
